+++ Project.toml
@@ -8,7 +8,6 @@ ADTypes = "47edcb42-4c32-4615-8424-f2b9edc5f35b"
 DocStringExtensions = "ffbed154-4ef7-542d-bbb7-c09d3a79fcae"
 FillArrays = "1a297f60-69ca-5386-bcde-b61e274b549b"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
-Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 
 [weakdeps]
 AbstractDifferentiation = "c29ec348-61ec-40c8-8164-b8c60e9d9f3d"
@@ -49,7 +48,6 @@ ForwardDiff = "0.10"
 JET = "0.8"
 LinearAlgebra = "1"
 PolyesterForwardDiff = "0.1"
-Random = "1"
 ReverseDiff = "1.15"
 Test = "1"
 Zygote = "0.6"
